evix: pass nix options through worker child env
Prepares generated `nix.conf` files on the parent side of `WorkerProcess`
spawning and passes `NIX_USER_CONF_FILES` with `Command::env` instead of
mutating process-global env. Arguably not the best, but it *is* the best
option I can think of.
