+++ crates/evix/src/worker_process.rs
@@ -1,4 +1,12 @@
-use std::{collections::VecDeque, env, path::Path, process::Stdio};
+use std::{
+  collections::VecDeque,
+  env,
+  ffi::{OsStr, OsString},
+  fs,
+  path::{Path, PathBuf},
+  process::{self, Stdio},
+  time::{SystemTime, UNIX_EPOCH},
+};
 
 use anyhow::{Context as _, Result, anyhow, bail};
 use tokio::{
@@ -25,11 +33,12 @@ const STDERR_READ_CHUNK: usize = 8 * 1024;
 const STDERR_LINE_LIMIT: usize = 16 * 1024;
 
 pub(crate) struct WorkerProcess {
-  pub(crate) label: String,
-  proc:             Child,
-  stdin:            Compat<ChildStdin>,
-  stdout:           Compat<BufReader<ChildStdout>>,
-  stderr_task:      JoinHandle<Result<String>>,
+  pub(crate) label:  String,
+  proc:              Child,
+  stdin:             Compat<ChildStdin>,
+  stdout:            Compat<BufReader<ChildStdout>>,
+  stderr_task:       JoinHandle<Result<String>>,
+  _nix_options_file: Option<NixOptionsFile>,
 }
 
 pub(crate) struct WorkResponse {
@@ -55,6 +64,7 @@ impl WorkerProcess {
       None => env::current_exe().context("resolving current exe")?,
     };
     debug!(worker = %label, "spawning local worker process");
+    let nix_options_file = prepare_nix_options_env(&config.nix_options)?;
     let mut command = Command::new(&exe);
     command
       .env(WORKER_ENV, "1")
@@ -62,6 +72,9 @@ impl WorkerProcess {
       .stdout(Stdio::piped())
       .stderr(Stdio::piped())
       .kill_on_drop(true);
+    if let Some(file) = &nix_options_file {
+      command.env("NIX_USER_CONF_FILES", file.conf_files());
+    }
 
     let mut child = command
       .spawn()
@@ -88,6 +101,7 @@ impl WorkerProcess {
       stdin,
       stdout,
       stderr_task,
+      _nix_options_file: nix_options_file,
     };
     worker.read_ready().await?;
     info!(worker = %worker.label, "worker ready");
@@ -196,6 +210,101 @@ impl WorkerProcess {
   }
 }
 
+/// Prepare caller-provided Nix settings for a worker subprocess.
+///
+/// These are evix's `--option KEY VALUE` pairs. The parent process writes the
+/// generated config and passes `NIX_USER_CONF_FILES` through the child's
+/// environment before the worker opens the Nix store or builds an eval state.
+fn prepare_nix_options_env(
+  options: &[(String, String)],
+) -> Result<Option<NixOptionsFile>> {
+  if options.is_empty() {
+    return Ok(None);
+  }
+
+  for (key, value) in options {
+    validate_nix_option_part("key", key)?;
+    validate_nix_option_part("value", value)?;
+  }
+
+  let path = env::temp_dir().join(format!(
+    "evix-nix-options-{}-{}.conf",
+    process::id(),
+    SystemTime::now()
+      .duration_since(UNIX_EPOCH)
+      .map_or(0, |duration| duration.as_nanos())
+  ));
+  let mut contents = String::new();
+  for (key, value) in options {
+    contents.push_str(key);
+    contents.push_str(" = ");
+    contents.push_str(value);
+    contents.push('\n');
+  }
+
+  fs::write(&path, contents).context("writing Nix options file")?;
+  let conf_files =
+    nix_user_conf_files(&path, env::var_os("NIX_USER_CONF_FILES").as_deref())?;
+  Ok(Some(NixOptionsFile { path, conf_files }))
+}
+
+#[derive(Debug)]
+struct NixOptionsFile {
+  path:       PathBuf,
+  conf_files: OsString,
+}
+
+impl NixOptionsFile {
+  fn conf_files(&self) -> &OsStr {
+    &self.conf_files
+  }
+}
+
+impl Drop for NixOptionsFile {
+  fn drop(&mut self) {
+    let _ = fs::remove_file(&self.path);
+  }
+}
+
+fn validate_nix_option_part(label: &str, part: &str) -> Result<()> {
+  if part.contains(['\n', '\r']) {
+    bail!("nix option {label} must not contain newlines");
+  }
+
+  match part.split_whitespace().next() {
+    Some("include" | "!include") => {
+      bail!("nix option {label} must not start with include directives")
+    },
+    _ => Ok(()),
+  }
+}
+
+fn nix_user_conf_files(
+  generated: &Path,
+  previous: Option<&OsStr>,
+) -> Result<OsString> {
+  let mut paths = vec![generated.to_path_buf()];
+
+  if let Some(previous) = previous.filter(|value| !value.is_empty()) {
+    paths.extend(env::split_paths(previous));
+  } else if let Some(default_config) = default_nix_user_conf_file() {
+    paths.push(default_config);
+  }
+
+  env::join_paths(paths).context("joining NIX_USER_CONF_FILES")
+}
+
+fn default_nix_user_conf_file() -> Option<PathBuf> {
+  let config_dir =
+    env::var_os("XDG_CONFIG_HOME")
+      .map(PathBuf::from)
+      .or_else(|| {
+        env::var_os("HOME").map(|home| PathBuf::from(home).join(".config"))
+      })?;
+  let config_file = config_dir.join("nix/nix.conf");
+  config_file.exists().then_some(config_file)
+}
+
 async fn capture_stderr<R>(label: String, mut stderr: R) -> Result<String>
 where
   R: AsyncRead + Unpin,
@@ -268,9 +377,47 @@ fn append_startup_hint(message: &mut String, phase: &str) {
 
 #[cfg(test)]
 mod tests {
+  use std::sync::Mutex;
+
   use tokio::io::AsyncWriteExt as _;
 
-  use super::{STDERR_TAIL_LIMIT, append_startup_hint, capture_stderr};
+  use super::*;
+
+  static ENV_LOCK: Mutex<()> = Mutex::new(());
+
+  struct EnvVarGuard {
+    name:     &'static str,
+    previous: Option<OsString>,
+  }
+
+  impl EnvVarGuard {
+    fn set(name: &'static str, value: Option<&OsStr>) -> Self {
+      let guard = Self {
+        name,
+        previous: env::var_os(name),
+      };
+      // SAFETY: tests using environment mutation hold ENV_LOCK.
+      unsafe {
+        match value {
+          Some(value) => env::set_var(name, value),
+          None => env::remove_var(name),
+        }
+      }
+      guard
+    }
+  }
+
+  impl Drop for EnvVarGuard {
+    fn drop(&mut self) {
+      // SAFETY: tests using environment mutation hold ENV_LOCK.
+      unsafe {
+        match &self.previous {
+          Some(value) => env::set_var(self.name, value),
+          None => env::remove_var(self.name),
+        }
+      }
+    }
+  }
 
   #[test]
   fn startup_hint_mentions_worker_dispatch() {
@@ -311,5 +458,78 @@ mod tests {
         assert_eq!(captured.len(), STDERR_TAIL_LIMIT);
         assert!(captured.bytes().all(|byte| byte == b'b'));
       });
+  }
+
+  #[test]
+  fn rejects_newlines_in_nix_option_parts() {
+    assert!(validate_nix_option_part("key", "foo\nbar").is_err());
+    assert!(validate_nix_option_part("value", "foo\rbar").is_err());
+  }
+
+  #[test]
+  fn rejects_include_directive_nix_option_parts() {
+    assert!(validate_nix_option_part("key", " include /tmp/nix.conf").is_err());
+    assert!(
+      validate_nix_option_part("value", "!include /tmp/nix.conf").is_err()
+    );
+    assert!(validate_nix_option_part("value", "allowed-uris").is_ok());
+  }
+
+  #[test]
+  fn nix_options_prepare_child_env_without_mutating_parent_env() {
+    let _lock = ENV_LOCK.lock().expect("env lock poisoned");
+    let previous = OsStr::new("/tmp/nix-a.conf:/tmp/nix-b.conf");
+    let _env_guard = EnvVarGuard::set("NIX_USER_CONF_FILES", Some(previous));
+
+    let options_file =
+      prepare_nix_options_env(&[("restrict-eval".into(), "true".into())])
+        .expect("nix options prepared")
+        .expect("options file created");
+    let options_path = options_file.path.clone();
+
+    let paths: Vec<_> = env::split_paths(options_file.conf_files()).collect();
+
+    assert_eq!(paths[0], options_path);
+    assert_eq!(paths[1], PathBuf::from("/tmp/nix-a.conf"));
+    assert_eq!(paths[2], PathBuf::from("/tmp/nix-b.conf"));
+    assert!(options_path.exists());
+    assert_eq!(env::var_os("NIX_USER_CONF_FILES"), Some(previous.into()));
+
+    drop(options_file);
+
+    assert!(!options_path.exists());
+    assert_eq!(env::var_os("NIX_USER_CONF_FILES"), Some(previous.into()));
+  }
+
+  #[test]
+  fn nix_options_include_default_user_conf_file_when_env_is_absent() {
+    let _lock = ENV_LOCK.lock().expect("env lock poisoned");
+    let test_dir = env::temp_dir().join(format!(
+      "evix-nix-options-test-{}-{}",
+      process::id(),
+      SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("system time after epoch")
+        .as_nanos()
+    ));
+    let config_dir = test_dir.join("xdg");
+    let default_config = config_dir.join("nix/nix.conf");
+    fs::create_dir_all(default_config.parent().expect("parent directory"))
+      .expect("create config dir");
+    fs::write(&default_config, "experimental-features = nix-command\n")
+      .expect("write default config");
+
+    let _nix_env = EnvVarGuard::set("NIX_USER_CONF_FILES", None);
+    let _xdg_env =
+      EnvVarGuard::set("XDG_CONFIG_HOME", Some(config_dir.as_os_str()));
+    let generated = test_dir.join("generated.conf");
+
+    let conf_files =
+      nix_user_conf_files(&generated, None).expect("conf files joined");
+    let paths: Vec<_> = env::split_paths(&conf_files).collect();
+
+    assert_eq!(paths, vec![generated, default_config]);
+
+    fs::remove_dir_all(test_dir).expect("remove test dir");
   }
 }
